refactored master
